Endrych/49 api pro editaci informaci uzivatele

+++ API/controllers/GameController.js
@@ -2,29 +2,34 @@ const Result = require('../models/Result');
 const ResultCodes = require('../enums/ResultCodes');
 const db = require('../config/dbconnection');
 const gameValidator = require('../validators/GameValidator');
+const processError = require('../helpers/processError');
+
 
 module.exports = app => {
 	function saveGameGenreGameTableData(id, data) {
 		return new Promise((resolve, reject) => {
 			var item = [];
-
-			data.forEach(element => {
-				item.push([id, element]);
-			});
-			db.query('DELETE FROM game_genre_games WHERE GameId = ?', [id], (err, res) => {
-				if (err) {
-					console.log(err);
-					throw err;
-				}
-
-				db.query('INSERT INTO game_genre_games (GameId,GameGenreId) VALUES ?', [item], (err, result) => {
-					if (err) {
-						throw err;
-					} else {
-						resolve();
-					}
+			if(data){
+				data.forEach(element => {
+					item.push([id, element]);
 				});
-			});
+				db.query('DELETE FROM game_genre_games WHERE GameId = ?', [id], (err, res) => {
+					if (err) {
+						console.log(err);
+						throw err;
+					}
+	
+					db.query('INSERT INTO game_genre_games (GameId,GameGenreId) VALUES ?', [item], (err, result) => {
+						if (err) {
+							throw err;
+						} else {
+							resolve();
+						}
+					});
+				});
+			}else{
+				resolve()
+			}
 		});
 	}
